nothing, still odd position errors, gona keep double anyway, feels better

+++ mygame/SphericalCoord.cs
@@ -9,10 +9,10 @@ namespace MyGame
 
     public struct SpehricalCoord
     {
-        public float longitude; // east to west
-        public float latitude; // top 0 bottom
-        public float altitude;
-        public SpehricalCoord(float longitude, float latitude, float altitude)
+        public double longitude; // east to west
+        public double latitude; // top 0 bottom
+        public double altitude;
+        public SpehricalCoord(double longitude, double latitude, double altitude)
         {
             this.longitude = longitude;
             this.latitude = latitude;
